refactor(#96) : DTO 분리
RequestDTO와 ResponseDTO들을 분리하였습니다.
