+++ src/main/java/odyssey/backend/node/controller/NodeController.java
@@ -5,8 +5,8 @@ import lombok.RequiredArgsConstructor;
 import odyssey.backend.global.response.CommonResponse;
 import odyssey.backend.global.response.ListCommonResponse;
 import odyssey.backend.global.response.SingleCommonResponse;
-import odyssey.backend.node.dto.NodeRequest;
-import odyssey.backend.node.dto.NodeResponse;
+import odyssey.backend.node.dto.request.NodeRequest;
+import odyssey.backend.node.dto.response.NodeResponse;
 import odyssey.backend.node.service.NodeService;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
