+++ src/main/java/odyssey/backend/node/dto/response/NodeResponse.java
@@ -1,4 +1,4 @@
-package odyssey.backend.node.dto;
+package odyssey.backend.node.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
