+++ src/main/java/odyssey/backend/node/service/NodeService.java
@@ -6,8 +6,8 @@ import odyssey.backend.node.exception.NodeNotFoundException;
 import odyssey.backend.roadmap.exception.RoadmapNotFoundException;
 import odyssey.backend.node.domain.Node;
 import odyssey.backend.node.domain.NodeRepository;
-import odyssey.backend.node.dto.NodeRequest;
-import odyssey.backend.node.dto.NodeResponse;
+import odyssey.backend.node.dto.request.NodeRequest;
+import odyssey.backend.node.dto.response.NodeResponse;
 import odyssey.backend.roadmap.domain.Roadmap;
 import odyssey.backend.roadmap.domain.RoadmapRepository;
 import org.springframework.stereotype.Service;
