+++ src/main/java/odyssey/backend/roadmap/controller/RoadmapController.java
@@ -5,10 +5,10 @@ import lombok.RequiredArgsConstructor;
 import odyssey.backend.global.response.CommonResponse;
 import odyssey.backend.global.response.ListCommonResponse;
 import odyssey.backend.global.response.SingleCommonResponse;
-import odyssey.backend.roadmap.dto.ImageUrlResponse;
-import odyssey.backend.roadmap.dto.RoadmapCountResponse;
-import odyssey.backend.roadmap.dto.RoadmapRequest;
-import odyssey.backend.roadmap.dto.RoadmapResponse;
+import odyssey.backend.roadmap.dto.response.ImageUrlResponse;
+import odyssey.backend.roadmap.dto.response.RoadmapCountResponse;
+import odyssey.backend.roadmap.dto.request.RoadmapRequest;
+import odyssey.backend.roadmap.dto.response.RoadmapResponse;
 import odyssey.backend.roadmap.service.RoadmapService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
