+++ src/main/java/odyssey/backend/roadmap/dto/response/RoadmapCountResponse.java
@@ -1,11 +1,11 @@
-package odyssey.backend.roadmap.dto;
+package odyssey.backend.roadmap.dto.response;
 
 import lombok.Getter;
 
 @Getter
 public class RoadmapCountResponse {
 
-    private Long count;
+    private final Long count;
 
     public RoadmapCountResponse(Long count) {
         this.count = count;
